move: 패키지 구조 변경에 따른 클래스 파일 이동 #3
